move EvaluateJS code into qwebelement_p
